Avoid avoidable warnings by explicit casts.

+++ test/test_chrono.cpp
@@ -252,7 +252,7 @@ private:
       tc::sleep(tc::milliseconds(25).convert<tc::nanoseconds>());
       tc::nanoseconds after = tc::now();
 
-      tc::default_repr_t diff = std::labs((after - now).raw() - 25000000);
+      tc::default_repr_t diff = std::labs(long((after - now).raw() - 25000000));
       CPPUNIT_ASSERT_MESSAGE("may fail if the system time changed", diff > 0);
       CPPUNIT_ASSERT_MESSAGE("may fail if there is high CPU load", diff < 200000);
     }
